grafico 1

+++ modal/reservamesa.php
@@ -14,12 +14,12 @@ class ReservaMesa{
         $this->camarero=$camarero;
         $this->dia=$dia;
         $this->hora=$hora;
-        $this->ocupacion=$ocupacion;    
+        $this->ocupacion=$ocupacion;
     }
 
     /**
      * Get the value of id
-     */ 
+     */
     public function getId()
     {
         return $this->id;
@@ -29,7 +29,7 @@ class ReservaMesa{
      * Set the value of id
      *
      * @return  self
-     */ 
+     */
     public function setId($id)
     {
         $this->id = $id;
@@ -39,7 +39,7 @@ class ReservaMesa{
 
     /**
      * Get the value of mesa
-     */ 
+     */
     public function getMesa()
     {
         return $this->mesa;
@@ -49,7 +49,7 @@ class ReservaMesa{
      * Set the value of mesa
      *
      * @return  self
-     */ 
+     */
     public function setMesa($mesa)
     {
         $this->mesa = $mesa;
@@ -59,7 +59,7 @@ class ReservaMesa{
 
     /**
      * Get the value of camarero
-     */ 
+     */
     public function getCamarero()
     {
         return $this->camarero;
@@ -69,7 +69,7 @@ class ReservaMesa{
      * Set the value of camarero
      *
      * @return  self
-     */ 
+     */
     public function setCamarero($camarero)
     {
         $this->camarero = $camarero;
@@ -79,7 +79,7 @@ class ReservaMesa{
 
     /**
      * Get the value of dia
-     */ 
+     */
     public function getDia()
     {
         return $this->dia;
@@ -89,7 +89,7 @@ class ReservaMesa{
      * Set the value of dia
      *
      * @return  self
-     */ 
+     */
     public function setDia($dia)
     {
         $this->dia = $dia;
@@ -99,7 +99,7 @@ class ReservaMesa{
 
     /**
      * Get the value of hora
-     */ 
+     */
     public function getHora()
     {
         return $this->hora;
@@ -109,7 +109,7 @@ class ReservaMesa{
      * Set the value of hora
      *
      * @return  self
-     */ 
+     */
     public function setHora($hora)
     {
         $this->hora = $hora;
@@ -119,7 +119,7 @@ class ReservaMesa{
 
     /**
      * Get the value of ocupacion
-     */ 
+     */
     public function getOcupacion()
     {
         return $this->ocupacion;
@@ -129,7 +129,7 @@ class ReservaMesa{
      * Set the value of ocupacion
      *
      * @return  self
-     */ 
+     */
     public function setOcupacion($ocupacion)
     {
         $this->ocupacion = $ocupacion;
@@ -143,7 +143,7 @@ class ReservaMesa{
         // echo "$alu->nombre";
         try {
             $stmt=$pdo->prepare("SELECT * FROM tbl_reserva_mesa rm  inner join tbl_mesa m on m.Id_mesa=rm.Id_mesa inner join tbl_sala s on s.Id_sala=m.Sala ORDER BY rm.Id_reserva DESC ");
-           /*  $stmt -> bindparam(  $stmt->bindParam(1,$id)); */
+            /*  $stmt -> bindparam(  $stmt->bindParam(1,$id)); */
             $stmt->execute();
             return $stmt;
         }catch (Exception $e){
@@ -153,38 +153,79 @@ class ReservaMesa{
 
     }
 
-    public static function getFilter($camarero,$sala,$mesa){
+    public static function getFilter($camarero,$sala,$mesa,$dia,$horaInicial,$horaFinal){
         require "../controller/conexion.php";
         try {
 
-            if($camarero==null && $sala==null && $mesa==null){
+            if($camarero==null && $sala==null && $mesa==null && $dia==null && $horaInicial==null && $horaFinal==null){
                 $sql="SELECT * FROM tbl_reserva_mesa rm  inner join tbl_mesa m on m.Id_mesa=rm.Id_mesa inner join tbl_sala s on s.Id_sala=m.Sala ORDER BY rm.Id_reserva DESC";
             }else{
                 $sql="SELECT * FROM tbl_reserva_mesa rm  inner join tbl_mesa m on m.Id_mesa=rm.Id_mesa inner join tbl_sala s on s.Id_sala=m.Sala ";
 
                 if($camarero==!null){
                     $sql=$sql." WHERE Id_cam LIKE '%$camarero%' ";
-                    if($sala==!null || $mesa==!null){
+                    if($sala==!null || $mesa==!null || $dia==!null || $horaInicial==!null || $horaFinal==!null){
                         $sql=$sql."AND";
                     }
                 }
                 if($sala==!null){
-                    
+
                     if($camarero==!null){
                         $sql=$sql." s.nombre_sala LIKE '%$sala%' ";
                     }else{
                         $sql=$sql." WHERE s.nombre_sala LIKE '%$sala%' ";
                     }
-                    if($mesa==!null){
+                    if($mesa==!null || $dia==!null || $horaInicial==!null || $horaFinal==!null){
                         $sql=$sql."AND";
                     }
                 }
                 if($mesa==!null){
-                    
+
                     if($sala==!null || $camarero==!null){
                         $sql=$sql." m.Id_mesa LIKE '%$mesa%' ";
                     }else{
                         $sql=$sql." WHERE m.Id_mesa LIKE '%$mesa%' ";
+                    }
+                    if($dia==!null || $horaInicial==!null || $horaFinal==!null){
+                        $sql=$sql."AND";
+                    }
+                }
+                if($dia==!null){
+                    /* $dia2=substr($dia,1);
+                    echo $dia2; */
+                    /* echo $dia; */
+                    echo $dia;
+
+                    if($sala==!null || $camarero==!null || $mesa==!null){
+                        $sql=$sql." rm.Dia_rm LIKE '$dia' ";
+                    }else{
+                        $sql=$sql." WHERE rm.Dia_rm LIKE '$dia' ";
+                    }
+                    if($horaInicial==!null || $horaFinal==!null){
+                        $sql=$sql."AND";
+                    }
+                }
+                if($horaInicial==!null || $horaFinal==!null){
+                    /* echo $horaInicial; */
+                    /* die(); */
+                    if($horaInicial==!null && $horaFinal==!null){
+                        if($sala==!null || $camarero==!null || $mesa==!null || $dia==!null){
+                            $sql=$sql." rm.Hora_rm BETWEEN '$horaInicial' AND '$horaFinal' ";
+                        }else{
+                            $sql=$sql." WHERE rm.Hora_rm BETWEEN '$horaInicial' AND '$horaFinal'";
+                        }
+                    }elseif($horaInicial==!null){
+                        if($sala==!null || $camarero==!null || $mesa==!null || $dia==!null){
+                            $sql=$sql." rm.Hora_rm > '$horaInicial' ";
+                        }else{
+                            $sql=$sql." WHERE rm.Hora_rm > '$horaInicial'";
+                        }
+                    }elseif($horaFinal==!null){
+                        if($sala==!null || $camarero==!null || $mesa==!null || $dia==!null){
+                            $sql=$sql." rm.Hora_rm < '$horaFinal' ";
+                        }else{
+                            $sql=$sql." WHERE rm.Hora_rm < '$horaFinal'";
+                        }
                     }
                 }
                 $sql=$sql." ORDER BY rm.Id_reserva DESC";
@@ -192,7 +233,7 @@ class ReservaMesa{
             /* echo $sql; */
             /* die(); */
             $stmt=$pdo->prepare($sql);
-           /*  $stmt -> bindparam(  $stmt->bindParam(1,$id)); */
+            /*  $stmt -> bindparam(  $stmt->bindParam(1,$id)); */
             $stmt->execute();
             return $stmt;
         }catch (Exception $e){
@@ -200,7 +241,18 @@ class ReservaMesa{
         }
     }
 
+    public function getMediasHora(int $sala){
+        require "../controller/conexion.php";
+
+        $sql="SELECT TIMEDIFF(rm.Hora_final_rm,rm.Hora_ini_rm) as 'mediaHoras', s.nombre_sala as nombre  FROM `tbl_reserva_mesa` rm INNER JOIN tbl_mesa m on rm.Id_mesa=m.Id_mesa INNER JOIN tbl_sala s on m.Sala=s.Id_sala GROUP by s.Id_sala ";
+        $stmt=$pdo->prepare($sql);
+        $stmt -> bindparam( 1,$sala);
+        $stmt->execute();
+
+        return $stmt;
+
+    }
     // public function subirRegistroMesa(){
-        
+
     // }
 }
